imporve cli_virtualbox

+++ cli_install_virtualbox_dmg.sh
@@ -13,36 +13,27 @@ cat ${LATEST_VIRTUAL_BOX_VERSION_FILE}
 
 LATEST_VIRTUAL_BOX_DMG_NAME="$(curl -s http://download.virtualbox.org/virtualbox/$(cat ${LATEST_VIRTUAL_BOX_VERSION_FILE})/ --list-only | grep dmg | sed -n 's/.*href="\([^"]*\).*/\1/p')"
 
-curl http://download.virtualbox.org/virtualbox/$(cat ${LATEST_VIRTUAL_BOX_VERSION_FILE})/${LATEST_VIRTUAL_BOX_DMG_NAME} -O
+# check download file already there
+if [[ -e ${LATEST_VIRTUAL_BOX_VERSION_FILE} ]]; then
+
+	echo " Ok we used file from directory"
+
+else
+
+	echo "Dowload file"
+	curl http://download.virtualbox.org/virtualbox/$(cat ${LATEST_VIRTUAL_BOX_VERSION_FILE})/${LATEST_VIRTUAL_BOX_DMG_NAME} -O
+
+fi
+
+echo "VirtualBox file => ${LATEST_VIRTUAL_BOX_DMG_NAME}"
 
 hdiutil mount ${LATEST_VIRTUAL_BOX_DMG_NAME}
 
 sudo installer -pkg /Volumes/VirtualBox/VirtualBox.pkg -target /
 
-hdiutil unmount /Volumes/VirtualBox/
+hdiutil unmount /Volumes/VirtualBox*
 
 # keep file
 # rm VirtualBox-4.2.12-84980-OSX.dmg
 
 exit 0
-
-# delete via cli
-# from here
-# https://github.com/iamrToday/pkg-remove/blob/master/pkg-remove.sh
-
-software="org.virtualbox.pkg.virtualbox"
-
-installPath=$(pkgutil --pkg-info ${software} | grep 'location:' | awk '{print $2}')
-cd /${installPath}
-#start uninstall
-pkgutil --only-files --files ${software} | tr '\n' '\0' | xargs -n 1 -0 rm -if
-pkgutil --only-dirs --files ${software} | tr '\n' '\0' | xargs -n 1 -0 rm -ifr
-sudo pkgutil --forget ${software}
-
-exit 0
-
-wget http://download.virtualbox.org/virtualbox/4.2.12/Oracle_VM_VirtualBox_Extension_Pack-4.2.12-84980.vbox-extpack
-$ sudo VBoxManage extpack install Oracle_VM_VirtualBox_Extension_Pack-4.2.12-84980.vbox-extpack
-$ rm Oracle_VM_VirtualBox_Extension_Pack-4.2.12-84980.vbox-extpack
-
-http://osxdaily.com/2014/07/31/manual-complete-app-removal-mac-os-x-terminal/
